chore: move theme checkout logic to composable
