+++ packages/default-theme/pages/checkout.vue
@@ -29,7 +29,10 @@
       <div class="checkout__aside desktop-only">
         <button @click="nextStep()">next</button>
         <transition name="fade">
-          <SidebarOrderSummary v-if="currentStep < CHECKOUT_STEPS.REVIEW" key="order-summary" />
+          <SidebarOrderSummary
+            v-if="currentStep < CHECKOUT_STEPS.REVIEW"
+            key="order-summary"
+          />
           <SidebarOrderReview
             v-else
             key="order-review"
@@ -48,11 +51,8 @@ import PaymentStep from '@shopware-pwa/default-theme/components/checkout/steps/P
 import PersonalDetailsStep from '@shopware-pwa/default-theme/components/checkout/steps/PersonalDetailsStep'
 import ShippingStep from '@shopware-pwa/default-theme/components/checkout/steps/ShippingStep'
 import OrderReviewStep from '@shopware-pwa/default-theme/components/checkout/steps/OrderReviewStep'
-import { useUser, useCheckout } from '@shopware-pwa/composables'
 import { ref, computed, reactive } from '@vue/composition-api'
-import { usePersonalDetailsStep } from '@shopware-pwa/default-theme/logic/checkout/usePersonalDetailsStep'
-import { useShippingStep } from '@shopware-pwa/default-theme/logic/checkout/useShippingStep'
-import { usePaymentStep } from '@shopware-pwa/default-theme/logic/checkout/usePaymentStep'
+import { useUICheckoutPage } from '@shopware-pwa/default-theme/logic/checkout/useUICheckoutPage'
 import {
   CHECKOUT_STEPS,
   getStepByNumber,
@@ -70,104 +70,12 @@ export default {
     SidebarOrderReview,
   },
   setup() {
-    const { isGuestOrder, createOrder } = useCheckout()
-    const {
-      isValid: isPersonalDetailsStepValid,
-      validate: validatePersonalDetailsStep,
-    } = usePersonalDetailsStep()
-    const {
-      isValid: isShippingStepValid,
-      validate: validateShippingStep,
-    } = useShippingStep()
-    const {
-      isValid: isPaymentStepValid,
-      validate: validatePaymentStep,
-    } = usePaymentStep()
-
-    const currentStep = ref(
-      isGuestOrder.value
-        ? CHECKOUT_STEPS.PERSONAL_DETAILS
-        : CHECKOUT_STEPS.REVIEW
-    )
-
-    const isPersonalDetailsStepCompleted = computed(() => {
-      return !isGuestOrder.value || isPersonalDetailsStepValid.value
-    })
-    const isShippingStepCompleted = computed(() => {
-      return !isGuestOrder.value || isShippingStepValid.value
-    })
-    const isPaymentStepCompleted = computed(() => {
-      return !isGuestOrder.value || isPaymentStepValid.value
-    })
-    const isReviewStepAvailable = computed(() => {
-      return !!isPaymentStepCompleted.value
-    })
-
-    const stepsStatus = computed(() => {
-      return {
-        PERSONAL_DETAILS: {
-          available: true,
-        },
-        SHIPPING: {
-          available: !!isPersonalDetailsStepCompleted.value,
-        },
-        PAYMENT: {
-          available: !!isShippingStepCompleted.value,
-        },
-        REVIEW: {
-          available: isReviewStepAvailable.value,
-        },
-      }
-    })
-
-    const nextStep = async (stepNumber) => {
-      let nextStepNumber = stepNumber || currentStep.value + 1
-      if (stepNumber === CHECKOUT_STEPS.PERSONAL_DETAILS)
-        nextStepNumber = CHECKOUT_STEPS.PERSONAL_DETAILS
-
-      if (
-        currentStep.value === CHECKOUT_STEPS.PERSONAL_DETAILS &&
-        currentStep.value !== nextStepNumber
-      )
-        validatePersonalDetailsStep()
-      if (
-        currentStep.value === CHECKOUT_STEPS.SHIPPING &&
-        currentStep.value !== nextStepNumber
-      )
-        validateShippingStep()
-      if (
-        currentStep.value === CHECKOUT_STEPS.PAYMENT &&
-        currentStep.value !== nextStepNumber
-      )
-        validatePaymentStep()
-
-      if (
-        currentStep.value === CHECKOUT_STEPS.REVIEW &&
-        nextStepNumber > CHECKOUT_STEPS.REVIEW
-      ) {
-        await createOrder()
-      } else {
-        const nextStep = getStepByNumber(nextStepNumber)
-        // console.error('STEP NO', nextStepNumber)
-        console.error('STEP NEXT', nextStep)
-        // console.error('NEXT STEP INVOKED', isPersonalDetailsStepValid.value)
-        console.error(
-          'Is NextStepAvailable',
-          stepsStatus.value[nextStep].available
-        )
-        if (stepsStatus.value[nextStep].available) {
-          currentStep.value = nextStepNumber
-          // this.$router.push({query: {step: nextStepNumber}})
-        }
-      }
-    }
+    const { currentStep, nextStep } = useUICheckoutPage()
 
     return {
       currentStep,
-      isGuestOrder,
-      isPersonalDetailsStepCompleted,
       nextStep,
-      CHECKOUT_STEPS
+      CHECKOUT_STEPS,
     }
   },
   watch: {
